[prod_env] update tests
